backend/requirements.txt
Flask==2.2.0
Werkzeug==2.2.2 # otherwise import error for flask
gunicorn==20.1.0
google-cloud-storage==1.43.0
oauth2client==4.1.3
google-cloud-logging==2.7.0
google-cloud-bigquery==3.25.0
# mistralai==0.0.11 
# google-cloud-tasks==2.7.1
google-api-python-client==2.34.0
google-auth==2.17.3
google-cloud-aiplatform==1.67.0
google-api-core==2.12.0
numpy
flask-cors